requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
class-registry==2.1.2
Django==4.2.20
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filters-django==1.0.5
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
idna==3.10
jsonpath==0.82.2
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
PyJWT==1.7.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
